Fixed import paths and moved all folders into items folder
